# fetch: improve performance of isValidHeaderValue